+++ gen_tests/security/lib/api.dart
@@ -1,0 +1,4 @@
+export 'package:security/api/default_api.dart';
+export 'package:security/api_client.dart';
+export 'package:security/api_exception.dart';
+export 'package:security/model/user.dart';